requirements.txt
# requirements.txt
# Dependencies pinned for stable deployment on free hosting environments
gradio>=4.44.0
pillow>=10.3.0
numpy<2.0
scikit-learn==1.4.2
opencv-python-headless>=4.8.1.78
pandas>=2.2.2